Move the setupStream functions into the main transcriber package
