+++ .gitignore
@@ -5,5 +5,13 @@
 .env*.local
 .env
 
+# Streams
+/data/backups/**/*
+/data/streams/**/*
+/data/uploads/**/*
+!/data/backups/.gitkeep
+!/data/streams/.gitkeep
+!/data/uploads/.gitkeep
+
 # Digital Ocean
 /for-server
