+++ packages/backend/package.json
@@ -11,10 +11,7 @@
     "start": "node dist/app.js"
   },
   "dependencies": {
-    "@dragongoose/streamlink": "^1.1.1",
     "@haishin/transcriber": "workspace:*",
-    "await-exec": "^0.1.2",
-    "date-fns": "^2.29.3",
     "express": "^4.18.2",
     "socket.io": "^4.6.1",
     "uuid": "^9.0.0"
