+++ packages/backend/src/handlers/stream-handler.ts
@@ -1,29 +1,5 @@
-import fs from 'fs'
-import exec from 'await-exec'
-import { Streamlink } from '@dragongoose/streamlink'
-import { format } from 'date-fns'
+import { setupStream } from "@haishin/transcriber"
 import transcriberHandler from "./transcriber-handler"
-
-interface StreamData {
-  file: string
-  newStream: boolean
-  originalUrl: string
-  streamUrl: string
-}
-
-function getPathsByUrl(url: string) {
-  const paths = url
-    .replace('http://', '')
-    .replace('https://', '')
-    .replace('c:', '')
-    .split('/')
-  const site = paths.shift()
-  
-  return {
-    site, 
-    user: paths.join('-')
-  }
-}
 
 function isUrl(str) {
   // Regular expression for URLs
@@ -31,72 +7,8 @@ function isUrl(str) {
   return urlRegex.test(str);
 }
 
-function setFileName(url: string): string {
-  const paths = getPathsByUrl(url);
-  const path = `${paths.site}--${paths.user}.mp4`;
-  return path;
-}
-
-function setArchivedFileName(url: string): string {
-  const dateTimeStart = format(new Date(), 'y-MM-dd_HH-mm-ss');
-  const paths = getPathsByUrl(url);
-  const path = `${paths.site}--${paths.user}--${dateTimeStart}.mp4`;
-  return path;
-}
-
-async function setupStream (originalUrl: string): Promise<StreamData> {
-  const file = `./data/streams/${setFileName(originalUrl)}`;
-
-  let {stdout: streamUrl} = await exec(`streamlink "${originalUrl}" best --stream-url`);
-  streamUrl = streamUrl.replace('\r\n', '');
-
-  try {
-    fs.accessSync(file, fs.constants.R_OK)
-    // If it exists then return the output
-    return {
-      newStream: false,
-      file,
-      originalUrl,
-      streamUrl,
-    };
-    // DO SOMETHING
-  } catch {
-    const client = new Streamlink(originalUrl, { outputStdout: true })
-    const streamFile = fs.createWriteStream(file)
-
-    client.begin()
-    client.on('log', data => {
-      // console.log('writing to file')
-      streamFile.write(data) // puts data into file
-    })
-
-    client.on('error', (error: Error) => {
-      console.log(error)
-    })
-  
-    client.on('close', () => {
-      streamFile.close() // closes the file when the stream ends or is closed
-  
-      // Move completed file to backups
-      fs.copyFileSync(file, `./data/backups/${setArchivedFileName(originalUrl)}`)
-
-      // Delete file in streams
-      fs.unlinkSync(file)
-    })
-
-    return {
-      newStream: true,
-      file,
-      originalUrl,
-      streamUrl,
-    };
-  }
-}
-
 export default function (io, socket) {
-  io.of('/').adapter.on('create-room', async (room) => {
-    console.log({ createRoom: room });
-    
+  io.of('/').adapter.on('create-room', async (room) => {    
     if (isUrl(room)) {
       // Get the initial stream data
       const streamData = await setupStream(room);
