+++ packages/backend/tsconfig.json
@@ -5,7 +5,7 @@
     "target": "es6",
     "moduleResolution": "node",
     "sourceMap": true,
-    "outDir": "dist"
+    "outDir": "dist",
   },
   "lib": ["es2015"],
   "include": ["**/*.ts"],
