+++ packages/frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@haishin/fe",
+  "name": "@haishin/frontend",
   "description": "A website that transcribes+translates Japanese streams into English with OpenAI+DeepL",
   "version": "2.0.0",
   "private": true,
