+++ packages/transcriber/src/utils/path-to-data.ts
@@ -1,0 +1,7 @@
+import path from 'path'
+
+const pathToData = (restOfFilePath: string): string => {
+  return path.join(__dirname, '../../../../', 'data', restOfFilePath);
+};
+
+export default pathToData;
