+++ packages/transcriber/tsconfig.json
@@ -5,7 +5,9 @@
     "target": "es6",
     "moduleResolution": "node",
     "sourceMap": true,
-    "outDir": "dist"
+    "outDir": "dist",
+    "declaration": true,
+    "declarationDir": "./types"
   },
   "lib": ["es2015"],
   "include": ["**/*.ts"],
