+++ packages/transcriber/types/externals.d.ts
@@ -1,2 +1,3 @@
 // Declaring modules that don't have typings here
+declare module 'await-exec';
 declare module 'node-ffprobe';
