+++ packages/transcriber/types/setup-stream.d.ts
@@ -1,0 +1,3 @@
+import type { StreamDataResponse } from '../types/responses.js';
+declare const setupStream: (originalUrl: string) => Promise<StreamDataResponse>;
+export default setupStream;
