+++ packages/transcriber/types/split-video-file.d.ts
@@ -1,0 +1,3 @@
+import { SplitVideoFileResponse } from '../types/responses';
+declare const splitVideoFile: (filename: string, startTime: number) => Promise<SplitVideoFileResponse>;
+export default splitVideoFile;
