+++ packages/transcriber/types/transcribe-translate-part.d.ts
@@ -1,0 +1,3 @@
+import { TranscriberResponse } from '../types/responses';
+declare const transcribeTranslatePart: (filename: string, prompt: string) => Promise<TranscriberResponse>;
+export default transcribeTranslatePart;
