+++ packages/transcriber/types/utils/path-to-data.d.ts
@@ -1,0 +1,2 @@
+declare const pathToData: (restOfFilePath: string) => string;
+export default pathToData;
